requirements.txt
acscore==0.3
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.11.1
djangorestframework==3.6.3
dulwich==0.17.3
kombu==4.0.2
pytz==2017.2
vine==1.1.3
